add different type

+++ storage/framework/views/d30ce9e36fe68606a9db5528c5b1d0b93a533458.php
@@ -67,33 +67,34 @@
                             </div>
                             <?php if(!empty($question['answers'])): ?>
                                 <?php $__currentLoopData = $question['answers']; $__env->addLoop($__currentLoopData); foreach($__currentLoopData as $answer): $__env->incrementLoopIndices(); $loop = $__env->getLastLoop(); ?>
-                                    <div id="collapse<?php echo e($question['id']); ?>" class="collapse show" aria-labelledby="heading<?php echo e($question['id']); ?>" data-parent="#accordion<?php echo e($question['id']); ?>">
-                                        <div class="card-body"  data-toggle="collapse">
-                                            <label>Ответ</label><div class="flex_answer js-edit-answer"><input data-answer="true" data-id="<?php echo e($answer['id']); ?>" data-name="answer" type="text" value="<?php echo e($answer['answer']); ?>" class="form-control inp"> <?php if(!empty($answer['branch_id'])): ?>  <a href="<?php echo e(SITE .  '/admin/selection/' . $question['selection_id'] . '/question/' . $answer['branch_id']); ?>"><button type="button" class="btn btn-secondary show_branch">Посмотреть ветку</button></a> <button type="button" data-id="<?php echo e($answer['id']); ?>" data-type="answer" class="btn btn-danger delete_branch">Удалить ветку</button> <?php else: ?>  <?php if(trim(strtolower($question['type'] === 'radio'))): ?> <a><button data-id="<?php echo e($answer['id']); ?>" data-selectionID="<?php echo e($selectionID); ?>" type="button" data-type="answer" class="btn btn-secondary add_branch_selection">Добавить ветку</button></a> <?php endif; ?> <?php endif; ?> <a><button type="button" data-id="<?php echo e($answer['id']); ?>" class="btn btn-danger delete_answer">Удалить ответ</button></a></div>
-                                        </div>
-                                        <br>
-                                        <div class="image_block_answer">
-                                            <form id="imageAnswer" class="form-group imageAnswerUpload" data-id="<?php echo e($answer['id']); ?>" method="POST" enctype="multipart/form-data" data-name="image_answer">
-                                                <button type="button"
-                                                        class="product-photos__item product-photos__item--add fas image_answer fa-plus"
-                                                        title="Крестик крутится, картинка грузится">
-                                                    <input type="file" class="button__input"
-                                                           data-answer="true"
-                                                           data-id="<?php echo e($answer['id']); ?>"
-                                                           id="imageAnswer"
-                                                           name="image_answer"
-                                                           accept=".jpg, .jpeg, .png, .webp, .heif"
-                                                           multiple>
-                                                </button>
-                                            </form>
-                                            <?php if(!empty($answer['img'])): ?>
-                                                <div class="image_anwer">
-                                                    <a href="<?php echo e($answer['img']); ?>" data-fancybox data-caption="<?php echo e($answer['answer']); ?>"><img style="width: 100px;height: 75px;" src="<?php echo e($answer['img']); ?>" class=""></a>
-                                                    <i class="fa fa-times deleteAnswerImage" onclick=" deleteAnswerImage('<?php echo e(SITE . '/admin/host/answer/deleteAnswerImage'); ?>')" data-id="<?php echo e($answer['id']); ?>" aria-hidden="true"></i>
-                                                </div>
-                                            <?php endif; ?>
-                                        </div>
-
+                                    <?php if(trim($question['type']) === 'radio' || trim($question['type']) === 'checkbox'): ?>
+                                        <div id="collapse<?php echo e($question['id']); ?>" class="collapse show" aria-labelledby="heading<?php echo e($question['id']); ?>" data-parent="#accordion<?php echo e($question['id']); ?>">
+                                            <div class="card-body"  data-toggle="collapse">
+                                                <label>Ответ</label><div class="flex_answer js-edit-answer"><input data-answer="true" data-id="<?php echo e($answer['id']); ?>" data-name="answer" type="text" value="<?php echo e($answer['answer']); ?>" class="form-control inp"> <?php if(!empty($answer['branch_id'])): ?>  <a href="<?php echo e(SITE .  '/admin/selection/' . $question['selection_id'] . '/question/' . $answer['branch_id']); ?>"><button type="button" class="btn btn-secondary show_branch">Посмотреть ветку</button></a> <button type="button" data-id="<?php echo e($answer['id']); ?>" data-type="answer" class="btn btn-danger delete_branch">Удалить ветку</button> <?php else: ?>  <?php if(trim(strtolower($question['type'] === 'radio'))): ?> <a><button data-id="<?php echo e($answer['id']); ?>" data-selectionID="<?php echo e($selectionID); ?>" type="button" data-type="answer" class="btn btn-secondary add_branch_selection">Добавить ветку</button></a> <?php endif; ?> <?php endif; ?> <a><button type="button" data-id="<?php echo e($answer['id']); ?>" class="btn btn-danger delete_answer">Удалить ответ</button></a></div>
+                                            </div>
+                                            <br>
+                                            <div class="image_block_answer">
+                                                <form id="imageAnswer" class="form-group imageAnswerUpload" data-id="<?php echo e($answer['id']); ?>" method="POST" enctype="multipart/form-data" data-name="image_answer">
+                                                    <button type="button"
+                                                            class="product-photos__item product-photos__item--add fas image_answer fa-plus"
+                                                            title="Крестик крутится, картинка грузится">
+                                                        <input type="file" class="button__input"
+                                                               data-answer="true"
+                                                               data-id="<?php echo e($answer['id']); ?>"
+                                                               id="imageAnswer"
+                                                               name="image_answer"
+                                                               accept=".jpg, .jpeg, .png, .webp, .heif"
+                                                               multiple>
+                                                    </button>
+                                                </form>
+                                                <?php if(!empty($answer['img'])): ?>
+                                                    <div class="image_anwer">
+                                                        <a href="<?php echo e($answer['img']); ?>" data-fancybox data-caption="<?php echo e($answer['answer']); ?>"><img style="width: 100px;height: 75px;" src="<?php echo e($answer['img']); ?>" class=""></a>
+                                                        <i class="fa fa-times deleteAnswerImage" onclick=" deleteAnswerImage('<?php echo e(SITE . '/admin/host/answer/deleteAnswerImage'); ?>')" data-id="<?php echo e($answer['id']); ?>" aria-hidden="true"></i>
+                                                    </div>
+                                                <?php endif; ?>
+                                            </div>
+                                        <?php endif; ?>
                                         <?php endforeach; $__env->popLoop(); $loop = $__env->getLastLoop(); ?>
                                         <?php else: ?>
                                             <?php if(!empty($question['answers'])): ?>
@@ -101,8 +102,60 @@
                                     </div>
                                     <?php endif; ?>
                                     <?php endif; ?>
-                                    <br>
-                                    <button type="button" data-id="<?php echo e($question['id']); ?>" class="btn btn-primary add_answer_selection">Добавить ответ</button>
+                                        <?php if(trim($question['type']) === 'select'): ?>
+                                            <?php if(!empty($question['answers'])): ?>
+                                                <div style="margin-left: 1%;margin-top: 1%;">
+                                                    <label>Значение в селекторе</label>
+                                                </div>
+                                                <?php $__currentLoopData = $question['answers']; $__env->addLoop($__currentLoopData); foreach($__currentLoopData as $answer): $__env->incrementLoopIndices(); $loop = $__env->getLastLoop(); ?>
+                                                    <div style="margin-bottom: -2%" class="card-body"  data-toggle="collapse">
+                                                        <div class="flex_answer js-edit-answer">
+                                                            <input data-answer="true" data-id="<?php echo e($answer['id']); ?>" data-name="answer" type="text" value="<?php echo e($answer['answer']); ?>" class="form-control inp"><button type="button" data-id="<?php echo e($answer['id']); ?>" class="btn btn-danger delete_answer">Удалить</button></a>
+                                                        </div>
+                                                    </div>
+                                                <?php endforeach; $__env->popLoop(); $loop = $__env->getLastLoop(); ?>
+                                            <?php endif; ?>
+                                            <br>
+                                            <button type="button" data-id="<?php echo e($question['id']); ?>" class="btn btn-primary add_answer_selection">Добавить</button>
+                                        <?php elseif(trim($question['type']) === 'range'): ?>
+                                            <div class="card-body">
+                                                <label>Минимальное значение</label>
+                                                <div class="flex_answer js-edit-answer">
+                                                    <input data-answer="true" data-id="<?php echo e($question['answers']['id'] ?? ''); ?>" data-name="min" type="text" value="<?php echo e($question['answers']['min'] ?? ''); ?>" class="form-control inp">
+                                                </div>
+                                                <label>Максимальное значение</label>
+                                                <div class="flex_answer js-edit-answer">
+                                                    <input data-answer="true" data-id="<?php echo e($question['answers']['id'] ?? ''); ?>" data-name="max" type="text" value="<?php echo e($question['answers']['max'] ?? ''); ?>" class="form-control inp">
+                                                </div>
+                                                <label>Начальное значение</label>
+                                                <div class="flex_answer js-edit-answer">
+                                                    <input data-answer="true" data-id="<?php echo e($question['answers']['id'] ?? ''); ?>" data-name="initial_value" type="text" value="<?php echo e($question['answers']['initial_value'] ?? ''); ?>" class="form-control inp">
+                                                </div>
+                                                <label>Шаг</label>
+                                                <div class="flex_answer js-edit-answer">
+                                                    <input data-answer="true" data-id="<?php echo e($question['answers']['id'] ?? ''); ?>" data-name="step" type="text" value="<?php echo e($question['answers']['step'] ?? ''); ?>" class="form-control inp">
+                                                </div>
+                                                <label>Шкала</label>
+                                                <div class="flex_answer js-edit-answer">
+                                                    <input data-answer="true" data-id="<?php echo e($question['answers']['id'] ?? ''); ?>" data-name="division" type="text" value="<?php echo e($question['answers']['division'] ?? ''); ?>" class="form-control inp">
+                                                </div>
+                                            </div>
+                                        <?php elseif(trim($question['type']) === 'textarea'): ?>
+                                            <div class="card-body">
+                                                <label>textarea</label>
+                                            </div>
+                                        <?php elseif(trim($question['type']) === 'image'): ?>
+                                            <div class="card-body">
+                                                <label>image</label>
+                                            </div>
+                                        <?php elseif(trim($question['type']) === 'file'): ?>
+                                            <div class="card-body">
+                                                <label>file</label>
+                                            </div>
+                                        <?php else: ?>
+                                            <br>
+                                            <button type="button" data-id="<?php echo e($question['id']); ?>" class="btn btn-primary add_answer_selection">Добавить ответ</button>
+                                        <?php endif; ?>
                         </div>
                     </div>
             <?php endif; ?>
